requirements.txt
apscheduler==3.6.3
Automat==0.8.0
chardet==3.0.4
docker==4.2.2
requests==2.20.0
pysyncobj>=0.2.3
subprocess32==3.5.4
scrapy>=2.0,<3.0
service-identity==18.1.0
six==1.15.0
sqlalchemy==1.3.18
sqlalchemy-migrate==0.13.0
sqlalchemy-utils
tornado==6.0.4
virtualenv>=20.0
w3lib==1.22.0
pyaml==20.4.0
pymysql==0.10.0
namegenerator==1.0.6
grpcio==1.31.0
grpcio-tools==1.31.0
